third_party/WebKit/LayoutTests/fast/dom/HTMLElement/attr-empty-string-expected.txt

When contentEditable attribute is empty string, element.contentEditable returns "true" and the element is editable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("div").getAttribute("contentEditable") is ""
PASS document.getElementById("div").contentEditable is "true"
PASS document.getElementById("div").isContentEditable is true
PASS window.getComputedStyle(div, "").getPropertyValue("-webkit-user-modify") is "read-write"
PASS successfullyParsed is true

TEST COMPLETE
